+++ AESCryptoTool/App.xaml.cs
@@ -2,7 +2,7 @@ using System.Windows;
 
 namespace AESCryptoTool
 {
-    public partial class App : Application
+    public partial class App : System.Windows.Application
     {
     }
 }